component seperated
